+++ README.rst
@@ -11,7 +11,7 @@
 
 
 ##############
-Nano-CAT 0.3.1
+Nano-CAT 0.3.2
 ##############
 
 **Nano-CAT** is a collection of tools for the analysis of nanocrystals,
