+++ nanoCAT/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '0.3.1'
+__version__ = '0.3.2'
